Module qrcode53bytes::info

Contains various QR specific encoding info.

Statics

FORMAT_INFO

Functions

block_data

block_ec_count
Returns error correction codewords per block.

format_info
Returns the format BitVec representation to be embedded.

group_block_count
Returns a vector of codewords counts per block. The length specifies how many blocks there are and each element how many codewords exist in that block.

total_bits
Returns the total data bits possible for a given version and error correction.

total_codewords
Returns the total codewords for a given version end error correction.